refactor: hoist POST memory-type and n-back scoring constants to server/lib leaves (#6818)
client/src/components/meatspace/post/constants.js hand-copied
MEMORY_DRILL_TYPES, COGNITIVE_DRILL_TYPES, COGNITIVE_LADDER_TYPES, and
nBackBalancedAccuracy from three different server modules with only
example-based tests as a guard. Hoist POST_SUPPORTED_MEMORY_TYPES out of
postValidation.js (which pulls zod) into the existing zod-free
postDrillTypes.js leaf, and nBackBalancedAccuracy out of
services/meatspacePost.js into a new postScoring.js leaf; the client now
re-exports postDrillTypes.js, postProgression.js (whose COGNITIVE_LADDER_TYPES
already existed there), and postScoring.js instead of carrying its own
copies. Every other export in constants.js (DOMAINS, DRILL_LABELS,
composedSessionDrillTypes, appliedNumeracyAnswerCorrect, ...) is unchanged.
